requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
idna==2.5
lxml==3.7.3
PyYAML==3.12
requests==2.18.1
six==1.10.0
slackclient==1.0.6
urllib3==1.21.1
websocket-client==0.44.0
